Split remove sequential blanks
